solution_iter

Function true_candidates

pub fn true_candidates<T: Solvable + BitOrAssign + Debug>(
    puzzle: &T,
    token: &CancellationToken,
) -> Option<T>

Compute all possible values that can be placed in any index for a puzzle.

BitOrAssign is used to combine multiple solutions and retain what possibilities are available at each index.

This uses a hybrid approach where a depth first search is attempted. If too many solutions are found, we switch to a breadth first approach and begin filling in possible values that have been skipped over so far.